requirements.txt
requests
websocket-client
pandas
python-dotenv
streamlit